moved validation rules yo helper
